Add JUnit5 System.exit() extension
As we will use exit codes in the code when the program gets invalid arguments
or fails to parse data for example, unit tests doesn't work properly when exit
calls are made.

This extension allows us to run tests with exit calls and also to test whether
exit codes returned are the ones we except.

It uses a deprecated way to allow these tests, SecurityManager. As there is no
alternative to SecurityManager which allows exit calls to be prevented, this
deprecated way is still used until a new Java version release changes allowing
detection of exit calls.

Source code of this extension is available on the following address:
https://github.com/tginsberg/junit5-system-exit