Add initial validation of arguments and show a help message
This validation shows a message in the case of error in the arguments
passed and exits with an error code 1.

A help message is shown in the case no arguments, -h or --help is entered.
An exit code 0 is returned in this case.

String resources have been added in a separate utility class, Messages.